refactor(types): Centralize type system and extract enums to constants
This commit implements a comprehensive type system refactoring that establishes
clear separation of concerns between type definitions and implementation code
across the entire codebase.

- **Centralized all enums** to `src/lib/constants/enums.ts`
  - AIProviderName, BedrockModels, OpenAIModels, VertexModels, GoogleAIModels
  - AnthropicModels, APIVersions, ErrorCategory, ErrorSeverity
- **Migrated HITL types** from `src/lib/hitl/types.ts` to `src/lib/types/hitlTypes.ts`
- **Extracted scattered types** to appropriate type modules:
  - Model registry types → `types/modelTypes.ts`
  - Provider health types → `types/providers.ts`
  - CLI types → `types/cli.ts`
  - Guardrails types → `types/guardrails.ts`
  - Utility types → `types/utilities.ts`

- **Consistent relative imports** throughout (`../constants/enums.js`)
- **Type-only imports** properly marked with `type` keyword
- **ESM compatibility** maintained with `.js` extensions

- Added `PerformanceMetrics` to analytics types
- Added `ContentFilteringResult` to guardrails types
- Added `ProviderRegistryOptions` and `ProviderHealthCheckOptions`
- Added `MCPStatus` and `ProviderStatus` to centralized locations
- Added `NeurolinkConstructorConfig` to config types
- Converted interfaces to type aliases for consistency
