+++ src/lib/hitl/index.ts
@@ -7,19 +7,6 @@
 // Core HITL Manager
 export { HITLManager } from "./hitlManager.js";
 
-// Type Definitions
-export type {
-  HITLConfig,
-  HITLRule,
-  ConfirmationRequest,
-  ConfirmationResult,
-  ConfirmationRequestEvent,
-  ConfirmationResponseEvent,
-  ConfirmationTimeoutEvent,
-  HITLStatistics,
-  HITLAuditLog,
-} from "./types.js";
-
 // Error Classes
 export {
   HITLError,
