+++ src/lib/providers/amazonBedrock.ts
@@ -20,10 +20,10 @@ import {
 } from "@aws-sdk/client-bedrock";
 import { BaseProvider } from "../core/baseProvider.js";
 import type {
-  AIProviderName,
   EnhancedGenerateResult,
   TextGenerationOptions,
 } from "../types/index.js";
+import { AIProviderName } from "../constants/enums.js";
 import type { StreamOptions, StreamResult } from "../types/streamTypes.js";
 import type { ToolDefinition, ToolArgs } from "../types/tools.js";
 import type { JsonValue } from "../types/common.js";
