+++ src/lib/providers/anthropicBaseProvider.ts
@@ -1,8 +1,7 @@
 import { createAnthropic } from "@ai-sdk/anthropic";
 import type { ZodType, ZodTypeDef } from "zod";
 import { streamText, type Schema, type LanguageModelV1 } from "ai";
-import type { AIProviderName } from "../types/index.js";
-import { AnthropicModels } from "../types/index.js";
+import { AIProviderName, AnthropicModels } from "../constants/enums.js";
 import type { StreamOptions, StreamResult } from "../types/streamTypes.js";
 import { BaseProvider } from "../core/baseProvider.js";
 import { logger } from "../utils/logger.js";
