+++ src/lib/providers/googleAiStudio.ts
@@ -1,8 +1,7 @@
 import { createGoogleGenerativeAI } from "@ai-sdk/google";
 import { streamText, type Schema, type LanguageModelV1 } from "ai";
 import type { ZodUnknownSchema } from "../types/typeAliases.js";
-import type { AIProviderName } from "../types/index.js";
-import { GoogleAIModels } from "../types/index.js";
+import { AIProviderName, GoogleAIModels } from "../constants/enums.js";
 import type {
   StreamOptions,
   StreamResult,
