+++ src/lib/types/analytics.ts
@@ -56,3 +56,17 @@ export type StreamAnalyticsData = {
   /** Provider analytics */
   providerAnalytics?: AnalyticsData;
 };
+
+export type PerformanceMetrics = {
+  startTime: number;
+  endTime?: number;
+  duration?: number;
+  memoryStart: NodeJS.MemoryUsage;
+  memoryEnd?: NodeJS.MemoryUsage;
+  memoryDelta?: {
+    rss: number;
+    heapTotal: number;
+    heapUsed: number;
+    external: number;
+  };
+};
