+++ src/lib/types/generateTypes.ts
@@ -4,7 +4,7 @@ import type {
   StandardRecord,
   ZodUnknownSchema,
 } from "./typeAliases.js";
-import type { AIProviderName } from "./providers.js";
+import { AIProviderName } from "../constants/enums.js";
 import type { AnalyticsData, TokenUsage } from "./analytics.js";
 import type { EvaluationData } from "./evaluation.js";
 import type { ChatMessage, ConversationMemoryConfig } from "./conversation.js";
